Add tree authorizations to fir_todos

+++ fir_todos/views.py
@@ -1,5 +1,6 @@
 import datetime
 
+from django.core.exceptions import PermissionDenied
 from django.dispatch import receiver
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.shortcuts import render, get_object_or_404
@@ -8,19 +9,25 @@ from django.http import HttpResponse
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.db.models import Q
 
-from incidents.views import is_incident_handler
-from incidents.models import Incident, model_created
+from incidents.authorization.decorator import authorization_required
+from incidents.views import is_incident_viewer
+from incidents.models import Incident, model_created, BusinessLine
 
 from fir_todos.models import TodoItem, TodoItemForm, TodoListTemplate
 
 
 @require_POST
 @login_required
-@user_passes_test(is_incident_handler)
-def create(request, incident_id):
-    incident = get_object_or_404(Incident, pk=incident_id)
+@authorization_required('incidents.handle_incidents', Incident, view_arg='incident_id')
+def create(request, incident_id, authorization_target=None):
+    if authorization_target is None:
+        incident = get_object_or_404(
+            Incident.authorization.for_user(request.user, 'incidents.handle_incidents'),
+            pk=incident_id)
+    else:
+        incident = authorization_target
 
-    form = TodoItemForm(request.POST)
+    form = TodoItemForm(request.POST, for_user=request.user)
     item = form.save(commit=False)
     item.incident = incident
     item.category = incident.category
@@ -31,12 +38,19 @@ def create(request, incident_id):
 
 
 @login_required
-@user_passes_test(is_incident_handler)
-def list(request, incident_id):
-    incident = get_object_or_404(Incident, pk=incident_id)
+@authorization_required('incidents.view_incidents', Incident, view_arg='incident_id')
+def list(request, incident_id, authorization_target=None):
+    if authorization_target is None:
+        incident = get_object_or_404(
+            Incident.authorization.for_user(request.user, 'incidents.handle_incidents'),
+            pk=incident_id)
+    else:
+        incident = authorization_target
     todos = incident.todoitem_set.all()
-    form = TodoItemForm()
-
+    if request.user.has_perm('incidents.handle_incidents', obj=incident):
+        form = TodoItemForm(for_user=request.user)
+    else:
+        form = None
     return render(
         request, 'fir_todos/list.html',
         {'todos': todos, 'form': form, 'incident_id': incident_id}
@@ -45,9 +59,10 @@ def list(request, incident_id):
 
 @require_POST
 @login_required
-@user_passes_test(is_incident_handler)
 def delete(request, todo_id):
     todo = get_object_or_404(TodoItem, pk=todo_id)
+    if not request.user.has_perm(todo.incident, 'incidents.handle_incidents'):
+        raise PermissionDenied()
     todo.delete()
 
     return HttpResponse('')
@@ -55,13 +70,14 @@ def delete(request, todo_id):
 
 @require_POST
 @login_required
-@user_passes_test(is_incident_handler)
 def toggle_status(request, todo_id):
     todo = get_object_or_404(TodoItem, pk=todo_id)
-    todo.done = not todo.done
-    if todo.done:
-        todo.done_time = datetime.datetime.now()
-    todo.save()
+    if (todo.business_line and request.user.has_perm('incidents.view_incidents', obj=todo.business_line)) or \
+            request.user.has_perm('incidents.view_incidents', obj=todo.incident):
+        todo.done = not todo.done
+        if todo.done:
+            todo.done_time = datetime.datetime.now()
+        todo.save()
 
     referer = request.META.get('HTTP_REFERER', None)
     dashboard = False
@@ -72,9 +88,10 @@ def toggle_status(request, todo_id):
 
 
 @login_required
-@user_passes_test(is_incident_handler)
+@user_passes_test(is_incident_viewer)
 def dashboard(request):
-    todos = TodoItem.objects.filter(business_line__name='CERT', incident__isnull=False, done=False)
+    bls = BusinessLine.authorization.for_user(request.user, 'incidents.view_incidents')
+    todos = TodoItem.objects.filter(business_line__in=bls, incident__isnull=False, done=False)
     todos = todos.select_related('incident', 'category')
     todos = todos.order_by('-incident__date')
